Use 64bit signed integer as file offset. <BTS:1247>
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@3672 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qm/src/model/account.cpp
@@ -1157,11 +1157,11 @@ qm::Account::Account(const WCHAR* pwszPath,
 	
 	Profile* pProfile = pImpl_->pProfile_;
 	
-	size_t nBlockSize = pProfile->getInt(L"Global", L"BlockSize");
+	unsigned int nBlockSize = pProfile->getInt(L"Global", L"BlockSize");
 	if (nBlockSize != -1)
 		nBlockSize *= 1024*1024;
 	
-	size_t nIndexBlockSize = pProfile->getInt(L"Global", L"IndexBlockSize");
+	unsigned int nIndexBlockSize = pProfile->getInt(L"Global", L"IndexBlockSize");
 	if (nIndexBlockSize == 0)
 		nIndexBlockSize = -1;
 	else if (nIndexBlockSize != -1)
